make the internals of coap_packet_t private
they are now only accessed using helper functions, for a package looks
different in lwip.

in addition, the local_interface is now passed to coap_handle_message
not as an argument (because that would mean mixing different arguments
in the population of the node's local interface), but as a part of the
incoming package (as assigned in coap_network_read).