Check if compiler supports -Wlogical-op
clang has dropped support for this warning and therefore emits
warnings when it is used as a compiler flag.

configure now uses AX_CHECK_COMPILE_FLAG[0] to check for compiler support.
[0] http://www.gnu.org/software/autoconf-archive/ax_check_compile_flag.html

Because local m4 macros are located in the sub-directory m4, autogen.sh
must not remove that directory but only clean the libtool-generated files.