doc/Makefile.am: fix up broken 'make dist'
The make target for creating the distribution tarball (make dist) was
broken some commits ago within the changes in the doc/ subdirectory. The
reason for this is that simple Doxygen has to be runs ones before the
automatism can pick up the files. automake uses dist_pkgdist_DATA as
trigger. So we have to use that variable.

Changing up the makefile template so the call for 'dist_pkgdist_DATA'
*and* 'all-local' is calling 'doxfile.stamp' so fist Doxygen is running.

By doing this also clean up a little bit and prepare the build for be
able to run out of the tree.