configure.ac: changing the platform related definitions
Simplifying the various definition for the currently supported
platforms. These are now Linux, Solaris and Darwin platforms. This will
hopefully grow up the next time but for now we only handle some extra
CFLAGS definitions and one extra define for Darwin. So there is not
really a need to handle this in two extra 'case switch' loops.

We have there some additional CFLAGS (right now only extra compiler
switches).  There is some "special" handling needed for getting this
info into the doxygen configuration. This switches are related on the
platform the build is running.