doc/Makefile.am: respect DESTDIR variable
The old behavior didn't allow a additonal setting of the DESTDIR
variable. This is a serious problem while building within a chroot for
example. Also the implementation of the stamp file was triggering
multiple builds in some cirumstances. This handling is reverted with the
remove of the old stamp file handling.