coap_io.c: select outbound interface when responding to multicast requests
coap_network_send() uses the address given by local_interface as
source address of outgoing IP packets. For responses to requests
that have been received via IP multicast, this address is a
multicast group and thus must not be set as source in the IP header.

As a workaround, this change clears the address and ifindex so that
the kernel selects a suitable outbound interface.