coap_io.c: avoid use of tainted data from recvmsg
coap_network_read() has used the address size reported by recvmsg()
in the msg_namelen field. As this value might be taken from the
incoming packet, coverity scan reports this value as a possible
security issue.

This fix sets the address size to the expected size and drops the
received packet if msg_namelen differs.