Makefile.am: Drop building coap hash functions
As "coap_key_t key" is no longer used as the key for hashing resources onto
the hashlist which gave the potential for coap hashes clashing for the
resource Uri, "str uri" being used instead, so disable the building of the
coap_hash functions.

In addition, the .sym and .map file information now has duplicate
entries removed.